Move auth outside to aumo/
